Introduce JdaService and delegate bot setup
