+++ src/main/java/ti4/commands/bothelper/BothelperCommand.java
@@ -4,7 +4,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.commands.CommandHelper;
 import ti4.commands.ParentCommand;
 import ti4.commands.Subcommand;
@@ -36,7 +36,7 @@ public class BothelperCommand implements ParentCommand {
     @Override
     public boolean accept(SlashCommandInteractionEvent event) {
         return ParentCommand.super.accept(event)
-                && CommandHelper.acceptIfHasRoles(event, AsyncTI4DiscordBot.bothelperRoles);
+                && CommandHelper.acceptIfHasRoles(event, JdaService.bothelperRoles);
     }
 
     @Override
