+++ src/main/java/ti4/commands/game/CreateGameButton.java
@@ -12,7 +12,7 @@ import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import net.dv8tion.jda.api.interactions.components.ActionRow;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.buttons.Buttons;
 import ti4.commands.CommandHelper;
 import ti4.commands.Subcommand;
@@ -74,7 +74,7 @@ class CreateGameButton extends Subcommand {
                     continue;
                 }
 
-                if (!member.getUser().isBot() && !CommandHelper.hasRole(event, AsyncTI4DiscordBot.developerRoles)) {
+                if (!member.getUser().isBot() && !CommandHelper.hasRole(event, JdaService.developerRoles)) {
                     int ongoingAmount = SearchGameHelper.searchGames(
                             member.getUser(), event, false, false, false, true, false, true, true, true);
                     int completedAndOngoingAmount = SearchGameHelper.searchGames(
