+++ src/main/java/ti4/helpers/DisasterWatchHelper.java
@@ -5,7 +5,7 @@ import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.utils.FileUpload;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.image.TileGenerator;
 import ti4.map.Game;
 import ti4.map.Tile;
@@ -17,13 +17,13 @@ public class DisasterWatchHelper {
 
     public static void postTileInDisasterWatch(
             Game game, GenericInteractionCreateEvent event, Tile tile, Integer rings, String message) {
-        if (AsyncTI4DiscordBot.guildPrimary
+        if (JdaService.guildPrimary
                         .getTextChannelsByName("disaster-watch-party", true)
                         .isEmpty()
                 || game.isFowMode()) {
             return;
         }
-        TextChannel watchParty = AsyncTI4DiscordBot.guildPrimary
+        TextChannel watchParty = JdaService.guildPrimary
                 .getTextChannelsByName("disaster-watch-party", true)
                 .getFirst();
         try (FileUpload systemWithContext =
@@ -35,13 +35,13 @@ public class DisasterWatchHelper {
     }
 
     public static void sendMessageInDisasterWatch(Game game, String message) {
-        if (AsyncTI4DiscordBot.guildPrimary
+        if (JdaService.guildPrimary
                         .getTextChannelsByName("disaster-watch-party", true)
                         .isEmpty()
                 || game.isFowMode()) {
             return;
         }
-        TextChannel watchParty = AsyncTI4DiscordBot.guildPrimary
+        TextChannel watchParty = JdaService.guildPrimary
                 .getTextChannelsByName("disaster-watch-party", true)
                 .getFirst();
         MessageHelper.sendMessageToChannel(watchParty, message);
