+++ src/main/java/ti4/helpers/Units.java
@@ -11,7 +11,7 @@ import java.util.regex.Pattern;
 import lombok.Data;
 import lombok.Getter;
 import org.jetbrains.annotations.Nullable;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.image.Mapper;
 import ti4.service.emoji.TI4Emoji;
 import ti4.service.emoji.UnitEmojis;
@@ -58,7 +58,7 @@ public class Units {
 
         @JsonIgnore
         public String getFileName() {
-            if (AsyncTI4DiscordBot.testingMode) return getFileName(false);
+            if (JdaService.testingMode) return getFileName(false);
             return getFileName(RandomHelper.isOneInX(Constants.EYE_CHANCE));
         }
