+++ src/main/java/ti4/listeners/MessageListener.java
@@ -12,7 +12,7 @@ import net.dv8tion.jda.api.entities.emoji.Emoji;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.apache.commons.lang3.StringUtils;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.executors.ExecutorServiceManager;
 import ti4.helpers.AliasHandler;
 import ti4.helpers.Constants;
@@ -39,8 +39,8 @@ public class MessageListener extends ListenerAdapter {
 
     @Override
     public void onMessageReceived(@Nonnull MessageReceivedEvent event) {
-        if (!AsyncTI4DiscordBot.isReadyToReceiveCommands()
-                || !AsyncTI4DiscordBot.isValidGuild(event.getGuild().getId())) {
+        if (!JdaService.isReadyToReceiveCommands()
+                || !JdaService.isValidGuild(event.getGuild().getId())) {
             return;
         }
 
@@ -69,7 +69,7 @@ public class MessageListener extends ListenerAdapter {
             }
             if ("572698679618568193".equalsIgnoreCase(event.getAuthor().getId())) {
                 TextChannel deletionLogChannel =
-                        AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("deletion-log", true).stream()
+                        JdaService.guildPrimary.getTextChannelsByName("deletion-log", true).stream()
                                 .findFirst()
                                 .orElse(null);
                 if (deletionLogChannel == null) return;
@@ -132,7 +132,7 @@ public class MessageListener extends ListenerAdapter {
         String msg2 = lfgRole.getAsMention()
                 + " this game is looking for more members (it's old if it has -launched [FULL] in its title) "
                 + message.getJumpUrl();
-        TextChannel lfgPings = AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("lfg-pings", true).stream()
+        TextChannel lfgPings = JdaService.guildPrimary.getTextChannelsByName("lfg-pings", true).stream()
                 .findFirst()
                 .orElse(null);
         MessageHelper.sendMessageToChannel(lfgPings, msg2);
@@ -337,15 +337,15 @@ public class MessageListener extends ListenerAdapter {
      * replicate messages in combat threads so that observers can see
      */
     private static void handleFogOfWarCombatThreadMirroring(MessageReceivedEvent event) {
-        if (!AsyncTI4DiscordBot.fowServers.isEmpty()
+        if (!JdaService.fowServers.isEmpty()
                 && // fog servers exists
-                !AsyncTI4DiscordBot.fowServers.contains(event.getGuild())
+                !JdaService.fowServers.contains(event.getGuild())
                 && // event server IS NOT the fog server
-                !AsyncTI4DiscordBot.guildCommunityPlays
+                !JdaService.guildCommunityPlays
                         .getId()
                         .equals(event.getGuild().getId())
                 && // NOR the community server
-                AsyncTI4DiscordBot.guildPrimaryID.equals(
+                JdaService.guildPrimaryID.equals(
                         Constants.ASYNCTI4_HUB_SERVER_ID)) { // bot is running in production
             return;
         } // else it's probably a dev/test server, so execute
