+++ src/main/java/ti4/map/Player.java
@@ -36,7 +36,7 @@ import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.buttons.Buttons;
 import ti4.draft.DraftBag;
 import ti4.draft.DraftItem;
@@ -289,7 +289,7 @@ public class Player extends PlayerProperties {
     @JsonIgnore
     public Role getRoleForCommunity() {
         try {
-            return AsyncTI4DiscordBot.jda.getRoleById(getRoleIDForCommunity());
+            return JdaService.jda.getRoleById(getRoleIDForCommunity());
         } catch (Exception e) {
         }
         return null;
@@ -299,7 +299,7 @@ public class Player extends PlayerProperties {
     @JsonIgnore
     public MessageChannel getPrivateChannel() {
         try {
-            return AsyncTI4DiscordBot.jda.getTextChannelById(getPrivateChannelID());
+            return JdaService.jda.getTextChannelById(getPrivateChannelID());
         } catch (Exception e) {
             return null;
         }
@@ -1152,14 +1152,14 @@ public class Player extends PlayerProperties {
 
     @JsonIgnore
     public User getUser() {
-        return AsyncTI4DiscordBot.jda.getUserById(getUserID());
+        return JdaService.jda.getUserById(getUserID());
     }
 
     @Override
     public String getUserName() {
         User userById = getUser();
         if (userById != null) {
-            Member member = AsyncTI4DiscordBot.guildPrimary.getMemberById(getUserID());
+            Member member = JdaService.guildPrimary.getMemberById(getUserID());
             if (member != null) {
                 setUserName(member.getEffectiveName());
             } else {
@@ -1235,7 +1235,7 @@ public class Player extends PlayerProperties {
             if (roleForCommunity == null && !getTeamMateIDs().isEmpty()) {
                 StringBuilder sb = new StringBuilder((noFactionIcon ? "" : getFactionEmoji()));
                 for (String userID : getTeamMateIDs()) {
-                    User userById = AsyncTI4DiscordBot.jda.getUserById(userID);
+                    User userById = JdaService.jda.getUserById(userID);
                     if (userById == null) {
                         continue;
                     }
