+++ src/main/java/ti4/service/game/CreateGameService.java
@@ -29,7 +29,7 @@ import net.dv8tion.jda.api.requests.restaction.ChannelAction;
 import net.dv8tion.jda.api.utils.FileUpload;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.ResourceHelper;
 import ti4.buttons.Buttons;
 import ti4.commands.CommandHelper;
@@ -71,7 +71,7 @@ public class CreateGameService {
         if (game == null) return;
 
         TextChannel bothelperLoungeChannel =
-                AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("staff-lounge", true).stream()
+                JdaService.guildPrimary.getTextChannelsByName("staff-lounge", true).stream()
                         .findFirst()
                         .orElse(null);
         if (bothelperLoungeChannel == null) return;
@@ -433,7 +433,7 @@ public class CreateGameService {
     }
 
     public static boolean gameOrRoleAlreadyExists(String name) {
-        List<Guild> guilds = AsyncTI4DiscordBot.jda.getGuilds();
+        List<Guild> guilds = JdaService.jda.getGuilds();
         List<String> gameAndRoleNames = new ArrayList<>();
 
         // GET ALL PBD ROLES FROM ALL GUILDS
@@ -452,7 +452,7 @@ public class CreateGameService {
     }
 
     private static List<Integer> getAllExistingPBDNumbers() {
-        List<Guild> guilds = new ArrayList<>(AsyncTI4DiscordBot.guilds);
+        List<Guild> guilds = new ArrayList<>(JdaService.guilds);
         List<Integer> pbdNumbers = new ArrayList<>();
 
         // GET ALL PBD ROLES FROM ALL GUILDS
@@ -486,13 +486,13 @@ public class CreateGameService {
 
     @Nullable
     private static Guild getServerWithMostCapacity() {
-        List<Guild> guilds = AsyncTI4DiscordBot.serversToCreateNewGamesOn.stream()
+        List<Guild> guilds = JdaService.serversToCreateNewGamesOn.stream()
                 .filter(CreateGameService::serverHasRoomForNewFullCategory)
                 .sorted(Comparator.comparing(CreateGameService::getServerCapacityForNewGames))
                 .toList();
 
-        if (guilds.isEmpty() && serverHasRoomForNewFullCategory(AsyncTI4DiscordBot.guildPrimary)) {
-            return AsyncTI4DiscordBot.guildPrimary;
+        if (guilds.isEmpty() && serverHasRoomForNewFullCategory(JdaService.guildPrimary)) {
+            return JdaService.guildPrimary;
         }
 
         if (guilds.isEmpty()) {
@@ -632,7 +632,7 @@ public class CreateGameService {
     }
 
     public static List<Category> getAllAvailablePBDCategories() {
-        return AsyncTI4DiscordBot.getAvailablePBDCategories();
+        return JdaService.getAvailablePBDCategories();
     }
 
     public static Category createNewCategory(String categoryName) {
@@ -643,7 +643,7 @@ public class CreateGameService {
             return null;
         }
 
-        List<Category> categories = AsyncTI4DiscordBot.jda.getCategoriesByName(categoryName, false);
+        List<Category> categories = JdaService.jda.getCategoriesByName(categoryName, false);
         if (!categories.isEmpty()) {
             String message = categories.stream().map(Channel::getAsMention).collect(Collectors.joining("\n"));
             BotLogger.info("Game Channel Creation - Category Already Exists:\n" + message);
@@ -685,7 +685,7 @@ public class CreateGameService {
     }
 
     public static boolean isLockedFromCreatingGames(GenericInteractionCreateEvent event) {
-        if (CommandHelper.hasRole(event, AsyncTI4DiscordBot.bothelperRoles)) {
+        if (CommandHelper.hasRole(event, JdaService.bothelperRoles)) {
             return false;
         }
