+++ src/main/java/ti4/service/game/EndGameService.java
@@ -17,7 +17,7 @@ import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.MessageChannel;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import org.apache.commons.lang3.StringUtils;
-import ti4.AsyncTI4DiscordBot;
+import ti4.JdaService;
 import ti4.helpers.Constants;
 import ti4.helpers.DisplayType;
 import ti4.helpers.Helper;
@@ -158,7 +158,7 @@ public class EndGameService {
 
         // GET BOTHELPER LOUNGE
         List<TextChannel> bothelperLoungeChannels =
-                AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("staff-lounge", true);
+                JdaService.guildPrimary.getTextChannelsByName("staff-lounge", true);
         TextChannel bothelperLoungeChannel =
                 !bothelperLoungeChannels.isEmpty() ? bothelperLoungeChannels.getFirst() : null;
         if (bothelperLoungeChannel != null) {
@@ -310,12 +310,12 @@ public class EndGameService {
 
     private static TextChannel getGameSummaryChannel(Game game) {
         List<TextChannel> textChannels;
-        if (game.isFowMode() && AsyncTI4DiscordBot.guildFogOfWar != null) {
-            ThreadArchiveHelper.checkThreadLimitAndArchive(AsyncTI4DiscordBot.guildFogOfWar);
-            textChannels = AsyncTI4DiscordBot.guildFogOfWar.getTextChannelsByName("fow-war-stories", true);
+        if (game.isFowMode() && JdaService.guildFogOfWar != null) {
+            ThreadArchiveHelper.checkThreadLimitAndArchive(JdaService.guildFogOfWar);
+            textChannels = JdaService.guildFogOfWar.getTextChannelsByName("fow-war-stories", true);
         } else {
-            ThreadArchiveHelper.checkThreadLimitAndArchive(AsyncTI4DiscordBot.guildPrimary);
-            textChannels = AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("the-pbd-chronicles", true);
+            ThreadArchiveHelper.checkThreadLimitAndArchive(JdaService.guildPrimary);
+            textChannels = JdaService.guildPrimary.getTextChannelsByName("the-pbd-chronicles", true);
         }
         return textChannels.isEmpty() ? null : textChannels.getFirst();
     }
